extract more types to ast-types
